requirements_frozen.txt
accelerate==1.11.0
annotated-doc==0.0.2
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
einops==0.8.1
fastapi==0.120.0
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==2.1.5
modelscope==1.31.0
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
packaging==25.0
pillow==11.3.0
protobuf==6.33.0
psutil==7.1.1
pydantic==2.12.3
pydantic_core==2.41.4
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
sentencepiece==0.2.1
setuptools==70.2.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.0+cpu
torchvision==0.24.0+cpu
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
